I change the value method, now the output array X has (only of stellar particles) positions, velocities, mass, normalized energy, circularity parameter and circularity parameter in the plane (in that order).
I also changed the decomposition methods to read the new input correctly.

Those particles that have positive or infinite energies, or circularity parameter greater than 1 and less than -1 have a label -1 in the Abadi method decomposition.

It is necessary to correct the other methods (so that the bad particles appear with a label -1).